Refactored Initial Resources data source code
- added placeholder for Hawkular implementation
- move InfluxDB implementation to a separate file
